tests(gateway): use fake_slurmweb_agent() more
